code/requirements.txt
asyncio==3.4.3
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
PyYAML==6.0
requests==2.28.0
trello.py==0.0.6.post4
py-trello==0.18.0
urllib3==1.26.9
discord.py==1.7.3
